refactored sidebar components
